requirements.clean.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.2.0
cycler==0.12.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
flatbuffers==25.2.10
fonttools==4.60.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scipy==1.11.4
six==1.17.0
sqlparse==0.5.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
